pyproject.toml
[project]
name = "neuromind"
version = "0.1.0"
description = "AI Assistant in Your Terminal"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langchain-ollama>=1.0.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.3",
    "jupyterlab>=4.4.10",
    "ipywidgets>=8.1.8",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.setuptools]
packages = ["agentor"]
